+++ .gitignore
@@ -1,5 +1,9 @@
-.pio
-.vscode/.browse.c_cpp.db*
-.vscode/c_cpp_properties.json
-.vscode/launch.json
-.vscode/ipch
+
+*.dblite
+*.bin
+*.elf
+*.o
+*.json
+*.a
+*.checksum
+*.tmp
