+++ README.md
@@ -1,3 +1,3 @@
-# Câmera Robô
-## Descrição
-- Software desenvolvido para controlar um robô que carrega uma câmera sobre um cabo. O movimento pelo cabo é feito por um motor de passo ligado a uma polia que gira uma roldana conectada ao cabo que está suspenso e preso em dois pontos, para garantir que os movimentos do motor sejam executados utilizaremos o sensor de infravermelho como feedback. 
+# IDRF
+
+software desenvolvido para controle de uma fechadura a solenoide que pode ser aberta de diversar formas inclusive via smartphone.
