+++ platformio.ini
@@ -21,3 +21,10 @@ framework = arduino
 lib_extra_dirs = ~/Documents/Arduino/libraries
 monitor_speed = 115200
 ; upload_port = Robo_Camera_1v0.local
+; build_flags = -DCORE_DEBUG_LEVEL=5
+; board_build.partitions = min_spiffs.csv
+; ; build_flags =
+; ;     -DBOARD_HAS_PSRAM
+; ;     -mfix-esp32-psram-cache-issue
+; upload_protocol = espota
+; upload_port = 192.168.0.172
