--- src/Strings.h
@@ -1,8 +1,0 @@
-#ifndef STRINGS_H
-#define STRINGS_H
-
-#include <Arduino.h>
-
-String
-
-#endif
