--- src/setupOTA.h
@@ -1,58 +1,0 @@
-#ifndef SETUPOTA_H
-#define SETUPOTA_H
-
-#include <Arduino.h>
-#include <ArduinoOTA.h>
-#include <globaldef.h>
-
-void Programming_OTA()
-{
-  String process = "| BOOT] ";
-  String step = "Setting > OVER_THE_AIR";
-  Serial.println("[" + String(millis()) + process + step);
-
-  ArduinoOTA.setHostname(DEVICE_NAME);
-  ArduinoOTA.onStart([]()
-                     {
-    String type;
-    if (ArduinoOTA.getCommand() == U_FLASH)
-    {
-      type = "sketch";
-    }
-    else
-    { // U_FS
-      type = "filesystem";
-    }
-
-    // NOTE: if updating FS this would be the place to unmount FS using FS.end()
-    Serial.println("Start updating " + type); });
-  ArduinoOTA.onEnd([]()
-                   { Serial.println("\nEnd"); });
-  ArduinoOTA.onProgress([](unsigned int progress, unsigned int total)
-                        { Serial.printf("Progress: %u%%\r", (progress / (total / 100))); });
-  ArduinoOTA.onError([](ota_error_t error)
-                     {
-    Serial.printf("Error[%u]: ", error);
-    if (error == OTA_AUTH_ERROR)
-    {
-      Serial.println("Auth Failed");
-    }
-    else if (error == OTA_BEGIN_ERROR)
-    {
-      Serial.println("Begin Failed");
-    }
-    else if (error == OTA_CONNECT_ERROR)
-    {
-      Serial.println("Connect Failed");
-    }
-    else if (error == OTA_RECEIVE_ERROR)
-    {
-      Serial.println("Receive Failed");
-    }
-    else if (error == OTA_END_ERROR)
-    {
-      Serial.println("End Failed");
-    } });
-  ArduinoOTA.begin();
-}
-#endif
